requirements.txt
tensorflow-cpu
flask
pillow
flask_cors
gunicorn
keras_preprocessing